engine: resources: Add Cloudflare DNS resource
This adds a resource to manage DNS on Cloudflare. Making use of the
official Golang API, we add functionality to create, update, edit,
delete records. A function to purge unmanaged records is also available.
Since the REST API does not provide us with an event stream, we make use
of polling to check for changes on Cloudflare.