lib: Automatically raise inotify limits at startup
It's not really dangerous to do this, and much nicer than letting the
user get an error. If someone knows why we wouldn't want to do this,
please let me know. Having said this, it would be nice to pool the
inotify watchers and maybe move to more efficient fanotify api's in the
future.