lang: funcs: operators: Catch invalid speculation
The below two tests cause speculation which crashes the type
unification, because it attempts to execute operator functions that
don't check that the type is valid. For speculation, we should probably
check this for safety, however it would be great if we could catch these
kinds of bugs in the general case for all functions.